Convert WebRTC streaming app to HLS video player
