+++ webrtc_streaming_test/pubspec.yaml
@@ -1,5 +1,5 @@
-name: webrtc_streaming_test
-description: "A new Flutter project."
+name: hls_video_player
+description: "HLS Video Player - Cross-platform streaming app"
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -35,10 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   
-  # RTMP streaming dependencies
-  camera: ^0.10.6
-  permission_handler: ^11.3.1
-  http: ^1.2.2
+  # HLS video player dependency
   video_player: ^2.8.2
 
 dev_dependencies:
